style抽离
